loki-kafka/requirements.txt
python-dotenv
websockets
kafka-python
requests